requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
prometheus-fastapi-instrumentator
python-json-logger
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi